added JWT authentication. Installed bcrypt and jsonwebtoken npms

+++ backend/controllers/goalController.js
@@ -1,16 +1,17 @@
 const asyncHandler = require("express-async-handler")
 
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
-// @desc       Get goals
+// @desc        Get goals
 // @route       GET /api/goals
 // @access      Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id })
 
     res.status(200).json(goals)
 })
-// @desc       Update goals
+// @desc        Update goals
 // @route       PUT /api/goals/:id
 // @access      Private
 const updateGoal =asyncHandler (async(req, res) =>{
@@ -20,12 +21,26 @@ const updateGoal =asyncHandler (async(req, res) =>{
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    //Checking to see if the logged in user matches the current user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal)
 })
-// @desc       Set goals
+
+// @desc        Set goals
 // @route       POST /api/goals/
 // @access      Private
 const setGoal =asyncHandler (async(req, res) =>{
@@ -35,10 +50,12 @@ const setGoal =asyncHandler (async(req, res) =>{
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
+
 // @desc       Delete goals
 // @route       DELETE /api/goals/:id
 // @access      Private
@@ -48,6 +65,19 @@ const deleteGoal =asyncHandler (async(req, res) =>{
     if(!goal){
         res.status(400)
         throw new Error("Can't find a goal with that id")
+    }
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    //Checking to see if the logged in user matches the current user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     await Goal.findByIdAndDelete(goal)
